Splits file-manager.js
